Третий этап модуляризации скуфрата
