--- modular_ss220/events/event_example/event_example.dm
@@ -1,7 +1,0 @@
-/datum/modpack/event_example
-	id = "event_example"
-	icon = 'modular_ss220/events/event_example/preview.dmi'
-	name = "event_example"
-	group = "Event"
-	desc = "Пример модпака для ивета."
-	author = "К*дер"
