--- modular_ss220/events/event_example/event_example.dme
@@ -1,8 +1,0 @@
-#ifndef EVENT_EXAMPLE
-#define EVENT_EXAMPLE
-
-#include "event_example.dm"
-
-#include "code\m_crayon.dm"
-
-#endif
