+++ modular_ss220/features/feature_example/feature_example.dm
@@ -5,46 +5,3 @@
 	group = "Features"
 	desc = "Пример написания модпаков для фич."
 	author = "К*дер"
-
-	update_data = list(
-		list(
-			"version" = "1.0",
-			"description" = "Initial release of the feature example modpack",
-			"date" = "2024-03-15"
-		),
-		list(
-			"version" = "1.1",
-			"description" = "Added new functionality and bug fixes",
-			"date" = "2024-04-01"
-		),
-		list(
-			"version" = "1.2",
-			"description" = "Optimized performance and added new features",
-			"date" = "2024-05-10"
-		),
-		list(
-			"version" = "1.3",
-			"description" = "И всё работает отлично!",
-			"date" = "2024-06-20"
-		),
-		list(
-			"version" = "1.4",
-			"description" = "updates overflow",
-			"date" = "2024-06-21"
-		),
-		list(
-			"version" = "1.5",
-			"description" = "updates overflow",
-			"date" = "2024-06-22"
-		),
-		list(
-			"version" = "1.6",
-			"description" = "updates overflow",
-			"date" = "2024-06-23"
-		),
-		list(
-			"version" = "1.7",
-			"description" = "updates overflow",
-			"date" = "2024-06-24"
-		)
-	)
