+++ modular_ss220/translations/translation_example/code/m_crayon.dm
@@ -1,3 +1,3 @@
 //ORIGINAL FILE: code/game/objects/effects/decals/crayon.dm
 /obj/effect/decal/cleanable/crayon
-	icon = 'modular_ss220/events/event_example/icons/crayondecal.dmi'
+	icon = 'modular_ss220/translations/translation_example/icons/crayondecal.dmi'
