+++ modular_ss220/translations/translation_example/icons/crayondecal.dmi
