+++ modular_ss220/translations/translation_example/translation_example.dm
@@ -1,0 +1,7 @@
+/datum/modpack/translation_example
+	id = "translation_example"
+	icon = 'modular_ss220/translations/translation_example/preview.dmi'
+	name = "translation_example"
+	group = "Translations"
+	desc = "Пример модпака для перевода."
+	author = "К*дер"
